作业6/requirements.txt
transformers>=4.30
torch>=2.0
requests>=2.28
Pillow>=9.0
